add parser folder

+++ Assets/Scripts/GameLogic/CardData.cs
@@ -6,7 +6,7 @@ public class CardData : ScriptableObject
 {
     public string cardName;
     public string description;
-    public int points;
+    public double points;
     public Sprite cardImage;
     public string imagePath;  // Ruta opcional, útil si necesitas más control
     public Player owner;
